requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
confluent-kafka==2.3.0
Django==5.0.3
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kafka-python==2.0.3.dev0
packaging==24.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.1
rpds-py==0.18.1
sqlparse==0.4.4
StrEnum==0.4.15
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
